+++ CHANGES.rst
@@ -90,7 +90,7 @@ New Features
 - Add Simplify button to subset plugin to make composite spectral subsets more user
   friendly. [#2237]
 
-- Plots within plugins can now be popped-out into their own windows. [#2254]
+- Plots within plugins can now be popped-out into their own windows. [#2254, #2314]
 
 - The ``specviz.load_spectrum`` method is deprecated; use ``specviz.load_data`` instead. [#2273]